Fix build and prefix with DCC
