+++ src/xcode/ENA/ENA/Source/AppDelegate & Globals/AppDelegate.swift
@@ -35,6 +35,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate, CoronaWarnAppDelegate, Re
 	override init() {
 		self.environmentProvider = Environments()
 
+		let some = DCCSignatureVerification()
+		let other = DCCSignatureVerifiableStub(error: nil)
+
 		#if DEBUG
 		if isUITesting {
 			self.store = MockTestStore()
